blkdev: print device size as 64-bit (was truncated to 32-bit)
blkdev_register() computed node.length in 64-bit but printed it with %u after
casting to uint32_t, so a device whose byte size is a multiple of 4 GiB printed
0 -- the 4 GiB root disk (8388608 * 512 = 0x1_0000_0000) logged "registered
(0 bytes)".  The stored length was always correct (df/statvfs read it fine);
only the boot message lied.  Print the full 64-bit value.

+++ sys/drivers/storage/blkdev.c
@@ -69,8 +69,8 @@ void blkdev_register(blkdev_t *dev) {
     dev->next = blkdev_list;
     blkdev_list = dev;
     
-    kprintf("Block device /dev/storage/%s registered (%u bytes)\n", 
-            dev->name, (uint32_t)dev->node.length);
+    kprintf("Block device /dev/storage/%s registered (%llu bytes)\n",
+            dev->name, (unsigned long long)dev->node.length);
 }
 
 void blkdev_unregister(blkdev_t *dev) {
